Register the inert driver
Simplify the calling conventions for inert by registering the port under
the atom 'inert'. All processes using the inert functions will use the
same instance of the driver:

    inert:start(),
    inert:poll(1).

The original interface has been moved to the prim_inert module. So if
several inert ports are needed:

    PollId = prim_inert:start(),
    prim_inert:poll(PollId, 1).

One issue is that the starting process owns the port but the registered
process is global. So if an unrelated process is using inert and the
owning process crashes, all processes using inert will crash.