poll/1,2: include the mode in the return value
Change the return value of poll/1,2 to a tagged tuple:

    {ok,read} | {ok,write} | {error,posix()}

Including the mode is required because polling an fd for read_write
notifications was ambiguous:

    ok = inert:poll(1) % same as inert:poll(1, [{mode,read}])
    ok = inert:poll(1, [{mode,read}]) % block until fd is ready for reading
    ok = inert:poll(1, [{mode,write}]) % block until fd is ready for writing

    % fd is ready for reading or writing?
    ok = inert:poll(1, [{mode,read_write}])

poll/1,2 will now always return the mode:

    {ok,write} = inert:poll(1, [{mode,read_write}])

The behaviour when poll uses read_write mode is still ambiguous and
needs further changes to be made deterministic. For example, when
poll/1,2 is called without a timeout, there is a race condition with the
fdclr that may leave a message in the mailbox (poll with timeout will
discard extra messages).