poll/1,2: flush mailbox for read_write
Since read and write are separate messages, it's possible that the
mailbox will have both if read_write mode is requested. There are a few
ways this could be handled:

* return the first received message, leave the other message queued

  A subsequent call to poll will return immediately.

* return the first received message, drop any other queued messages

  This is the way poll with read_write mode acts today when a timeout is
  hit.

* check for both messages and, if present, return read_write.

Option 1 is racy: Who knows what will happen to the fd between one call
to poll and the next. But it should be harmless since poll does not act
directly on the fd, i.e., the library that directly manipulates the fd
will receive the error (probably {error,eagain}).

Option 3 might be the way to go but adds some overhead:

    case inert:poll(1, [{mode,read_write}]) of
        {ok, read} -> do_read(1);
        {ok, write} -> do_write(1);
        {ok, read_write} -> do_read(1), do_write(1)
    end.

Since option 2 is the way poll handles timeouts today, re-use the same
code for the non-timeout case.