Fix Makefile $(REBAR) variable handling on Illumos
On solaris flavors, `which` has sometimes poor behavior of
not returning proper exit codes or printing errors to STDOUT
rather than STDERR.  All of this is to say, the makefile didn't
behave correctly on Solaris. $(REBAR) ended up being set to
"rebar not found" which made the compile target less than happy.