Use C99 types
'u_int32_t' is not a standard type in c, but it is defined on several
platforms in 'sys/types.h'. However, inert doesn't include that
header. Instead, inert has relied on transitively-including it via
'erl_drriver.h' and 'ei.h'. Something changed in OTP23, breaking that
transitive include on some Linux distros.